+++ packages/core/src/__test__/client.test.ts
@@ -895,7 +895,6 @@ describe("client", () => {
 
       expect(response).toEqual({
         txHash: "tx_hash123",
-        success: true,
       });
     });
   });
@@ -1636,18 +1635,6 @@ describe("client", () => {
       const result = await client.getRecommendedGasPrice("noble-1");
 
       expect(result?.denom).toEqual("uusdc");
-    });
-
-    it("returns the recommended gas price for Quasar (is IBC token)", async () => {
-      const client = new SkipRouter({
-        apiURL: SKIP_API_URL,
-      });
-
-      const result = await client.getRecommendedGasPrice("quasar-1");
-
-      expect(result?.denom).toEqual(
-        "ibc/0471F1C4E7AFD3F07702BEF6DC365268D64570F7C1FDC98EA6098DD6DE59817B",
-      );
     });
   });