+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @skip-router/core
 
+## 2.0.4
+
+### Patch Changes
+
+- 8e0de72: bump cosmjs version and chain registry
+
+## 2.0.3
+
+### Patch Changes
+
+- 4903d27: executeSVMTransaction api submit after sign
+
 ## 2.0.2
 
 ### Patch Changes