+++ packages/examples/CHANGELOG.md
@@ -1,5 +1,19 @@
 # examples
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [8e0de72]
+  - @skip-router/core@2.0.4
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [4903d27]
+  - @skip-router/core@2.0.3
+
 ## 1.0.3
 
 ### Patch Changes